+++ src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,5 +1,5 @@
 import { Component, type ReactNode } from "react";
-import { Button } from "../Button";
+import { Button } from "@/components";
 
 type Props = {
   children: ReactNode;
