+++ src/components/ErrorBoundaryTestButton/ErrorBoundaryTestButton.tsx
@@ -1,5 +1,5 @@
 import { Component, type ReactNode } from "react";
-import { Button } from "../Button";
+import { Button } from "@/components";
 
 type State = {
   brokeRender: boolean;
