+++ src/components/PeopleList/PeopleList.tsx
@@ -1,5 +1,5 @@
 import { Component, type ReactNode } from "react";
-import type { People } from "../../types/People";
+import type { People } from "@/types/People";
 import PeopleListItem from "./PeopleListItem";
 
 type Props = {
