+++ src/components/index.ts
@@ -4,3 +4,4 @@ export * from "./PeopleList";
 export * from "./Loading";
 export * from "./Alert";
 export * from "./ErrorBoundaryTestButton";
+export * from "./ErrorBoundary";
