+++ src/main.tsx
@@ -1,8 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./main.css";
+
 import App from "./App";
-import { ErrorBoundary } from "./components/ErrorBoundary";
+import { ErrorBoundary } from "@/components";
+
+import "./main.css";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
